Use pip to install stacktach packages.
Use "pip install -e ." rather than "python setup.py install"
to install the stacktach packages for the sandbox.

The setup.py invokes easy_install, which does not work well with
pbr, which we use for requirements. This can cause install of stacktach
requirements to be skipped on a sandbox build. Using pip fixes that.

Change-Id: Ia376806c315a156b39aef9fa21d9ccac6e01e6c4